refactor: mover lógica de validación y actualización de imagenes a ImageFacade
